+++ src/pages/home/home sections/BlogSection/BlogPostPreview.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { RichTextDisplay } from "../../../../components/RichTextEditor/RichTextDisplay";
 import { Row } from "../../../../components/Row";
 import { BlogPostProps } from "../../../admin/Blog/blogPostProps";
-import { fancyDisplayTimestamp } from "../../../blog/BlogPost/BlogPost";
+import { fancyDisplayTimestamp } from "../../../blog/BlogPost/BlogPost/BlogPost";
 
 const BlogPostPreviewContainer = styled(motion.button)`
   position: relative;
